# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Social Network",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://social.loc/api"
        }
    ],
    "paths": {
...
        "/users/{user}/posts": {
            "get": {
                "tags": [
                    "feed"
                ],
                "summary": "Get user's feed posts",
                "operationId": "App\\Http\\Controllers\\PostController::index",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "feed"
                ],
                "summary": "Post to user's feed",
                "operationId": "App\\Http\\Controllers\\PostController::store",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "This user is not in friends list"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "This action is unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "PostCollection": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Post"
                        }
                    },
                    "links": {
                        "type": "object",
                        "example": {
                            "first": "first page link",
                            "last": "last page link",
                            "next": "next page link",
                            "prev": "prev page link"
                        }
                    },
                    "meta": {
                        "type": "object",
                        "example": {
                            "current_page": 1,
                            "from": 1,
                            "last_page": 1,
                            "path": "full url path",
                            "per_page": 10,
                            "to": 2,
                            "total": 2
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "PostCollection"
                }
            },
            "PostResource": {
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/Post"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "PostResource"
                }
            },
...
            "Post": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "user_id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "posted_user_id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "2"
                    },
                    "posted_user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "is_public": {
                        "type": "boolean",
                        "readOnly": "true",
                        "example": "true"
                    },
                    "text": {
                        "type": "string",
                        "example": "Post text"
                    },
                    "created_at": {
                        "description": "Registration date and time",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true",
                        "example": "2021-01-15 12:59:20"
                    },
                    "updated_at": {
                        "description": "Last updated date and time",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true",
                        "example": "2021-01-20 04:25:15"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Post"
                }
            },
...
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "John"
                    },
                    "surname": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Doe"
                    },
                    "avatar": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "http://site.com/img/avatar.png"
                    },
                    "friends_count": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "355"
                    },
                    "email": {
                        "description": "User unique email address",
                        "type": "string",
                        "format": "email",
                        "readOnly": "true",
                        "example": "user@gmail.com"
                    },
                    "created_at": {
                        "description": "Registration date and time",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true",
                        "example": "2021-01-15 12:59:20"
                    },
                    "updated_at": {
                        "description": "Last updated date and time",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true",
                        "example": "2021-01-20 04:25:15"
                    },
                    "join_date": {
                        "description": "Registration date in human radable format",
                        "type": "string",
                        "readOnly": "true",
                        "example": "10 days ago"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            }
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Use /login endpoint to get access token",
                "name": "Bearer",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}